+++ Yosemite/Yosemite/Model/Storage/OrderCount+ReadOnlyConvertible.swift
@@ -9,7 +9,7 @@ extension Storage.OrderCount: ReadOnlyConvertible {
     /// Updates the Storage.OrderCount with the ReadOnly.
     ///
     public func update(with orderCount: Yosemite.OrderCount) {
-        siteID = Int64(orderCount.siteID)
+        siteID = orderCount.siteID
     }
 
     /// Returns a ReadOnly version of the receiver.
@@ -17,7 +17,7 @@ extension Storage.OrderCount: ReadOnlyConvertible {
     public func toReadOnly() -> Yosemite.OrderCount {
         let orderCountItems = items?.map { $0.toReadOnly() } ?? [Yosemite.OrderCountItem]()
 
-        return OrderCount(siteID: Int64(siteID),
+        return OrderCount(siteID: siteID,
                           items: orderCountItems)
     }
 }
