+++ Yosemite/Yosemite/Model/Storage/OrderCoupon+ReadOnlyConvertible.swift
@@ -9,7 +9,7 @@ extension Storage.OrderCoupon: ReadOnlyConvertible {
     /// Updates the Storage.OrderCoupon with the ReadOnly.
     ///
     public func update(with orderCoupon: Yosemite.OrderCouponLine) {
-        couponID = Int64(orderCoupon.couponID)
+        couponID = orderCoupon.couponID
         code = orderCoupon.code
         discount = orderCoupon.discount
         discountTax = orderCoupon.discountTax
@@ -18,7 +18,7 @@ extension Storage.OrderCoupon: ReadOnlyConvertible {
     /// Returns a ReadOnly version of the receiver.
     ///
     public func toReadOnly() -> Yosemite.OrderCouponLine {
-        return OrderCouponLine(couponID: Int64(couponID),
+        return OrderCouponLine(couponID: couponID,
                                code: code ?? "",
                                discount: discount ?? "",
                                discountTax: discountTax ?? "")
