+++ Yosemite/Yosemite/Model/Storage/OrderStatus+ReadOnlyConvertible.swift
@@ -10,7 +10,7 @@ extension Storage.OrderStatus: ReadOnlyConvertible {
     ///
     public func update(with orderStatus: Yosemite.OrderStatus) {
         name = orderStatus.name
-        siteID = Int64(orderStatus.siteID)
+        siteID = orderStatus.siteID
         slug = orderStatus.slug
         total = Int64(orderStatus.total)
     }
@@ -18,6 +18,6 @@ extension Storage.OrderStatus: ReadOnlyConvertible {
     /// Returns a ReadOnly version of the receiver.
     ///
     public func toReadOnly() -> Yosemite.OrderStatus {
-        return OrderStatus(name: name, siteID: Int64(siteID), slug: slug, total: Int(total))
+        return OrderStatus(name: name, siteID: siteID, slug: slug, total: Int(total))
     }
 }
